added js confirm

+++ admin/admin-manage.php
@@ -1,6 +1,18 @@
+<?php 
+$user_id = $_SESSION['user_id'];
+//delete parser
+if( $_GET['action'] == 'delete' ){
+	//which post are we deleting?
+	$post_id = $_GET['post_id'];
+	$query_delete = "DELETE FROM posts
+				WHERE post_id = $post_id
+				AND user_id = $user_id";
+	$result_query  = $db->query($query_delete);	
+} ?>
+
 <h2>Manage Your Posts</h2>
 <?php //get all the posts that were written by the logged in user
-$user_id = $_SESSION['user_id'];
+
 $query = "SELECT post_id, title, is_published
 	    FROM posts
 	    WHERE user_id = $user_id
@@ -15,10 +27,22 @@ if( $result->num_rows >= 1 ){ ?>
 		<a href="<?php echo SITE_URL; ?>/admin/?page=edit&amp;post_id=<?php echo $row['post_id'] ?>">
 			edit
 		</a>
-		<a href="?action=delete&amp;post_id=<?php echo $row['post_id'] ?>" class="warn">
+		<a href="?page=manage&amp;action=delete&amp;post_id=<?php echo $row['post_id'] ?>" class="warn" onclick="return confirmAction()">
 			delete
 		</a>
 	</li>
 	<?php } ?>
 </ul>
+
+<script>
+// Function to confirm permanent actions. use onclick.
+function confirmAction(){
+	var agree = confirm("This is permanent. Are you sure?");
+	if(agree){
+		return true;
+	}else{
+		return false;
+	}
+}
+</script>
 <?php }//end if there are posts ?>
